♲ Minor refactorings in core
