+++ core/src/index.ts
@@ -1,6 +1,8 @@
-export { default } from './cuillere'
+export { cuillere as default } from './cuillere'
 export * from './cuillere'
+export * from './executable-promise'
 export * from './generator'
 export * from './plugins'
 export * from './operations'
 export { Task } from './stack'
+export * from './time'
