--- core/src/utils/ts-generator.ts
@@ -1,2 +1,0 @@
-export type TSGenerator<T = unknown, R = any, N = unknown> = Generator<T, R, N>
-export type TSAsyncGenerator<T = unknown, R = any, N = unknown> = AsyncGenerator<T, R, N>
